Updated openstack/openstack Project: openstack/neutron  b572eb233cb76aeda3419080459aaa1bbbc409b2
sanity check: deprecate all version based checks

Those are against the spirit of the tool, and were introduced by
mistake. Sanity checks should validate actual features not version
numbers, because distributions can ship an older version of a package
that still has the needed patches backported to support features needed
for neutron; in which case the sanity check tool should not fail on
those deployments.

Updated 'effective neutron' guide to reflect the unspoken rule.

Since the tool is supposed to be consumed by operators, and not (just)
developers, we can't remove those CLI arguments immediately, and so we
follow deprecation process instead. In Queens, we will completely remove
those checks and options.

Change-Id: I2df8b7b39bca6d548a6f17466c26ffa688eb1d59