Updated openstack/openstack Project: openstack/cinder  ed8e5f80c5803b12f8add072f3b556a8525ee243
RemoteFS: fix volume cascade delete

Attempting to delete a volume along with its snapshots fails for
RemoteFS based drivers. The reason is that that the driver expects
the volume status to be 'available', 'in-use' or 'backing-up',
while in this situation the volume status would be 'deleting'.

Most of those drivers apply locks to the volume/snapshot deletion
methods, so it should be safe to just add 'deleting' to the list of
acceptable volume states for this operation.

Change-Id: I1242d03fb6bbbd40ac2dace1aa865a6f91c22d23
Closes-Bug: #1681406