Updated openstack/openstack Project: openstack-infra/tripleo-ci  493fba496886845eca15a888597bd1a701ffbc66
Always rebuild docs venv for tripleo-docs

If we continue to reuse the previous venv we don't pick up new
changes in tripleosphinx because it's installed with
python setup.py install, which apparently doesn't update the code
if it's already installed.

pip should have cached all of the dependencies, so this will mostly
just drive a little more CPU on the website vm.  It's not under
significant load though, so the extra is unlikely to be noticeable.

Change-Id: I38649e4288146edc679c19018755a5296d3b8031