Updated openstack/openstack Project: openstack/neutron  6cffc6cc68219029c700ec53dbb836bd7547687f
docs: Update TESTING.rst about openvswitch requirements

Since we switched to use Ubuntu Cloud Archives in upstream gate, the
note in TESTING.rst is not longer valid for the gate. The condition to
trigger the test still remains in order to avoid false negative test
results for those that run functional tests with older openvswitch
version.

Change-Id: Ibff983efe0f452864d003913205cbbfa39ee7374