Updated openstack/openstack Project: openstack/heat  7dacf57a9a52fd9df8564c62724639e5100be5c5
Wait for the policy to be done in tests

As observed in the amqp job, we sometimes signal scale policy too fast,
without waiting for the medata to be set. It creates a timeout. We can
check the metadata of the policy before signaling again, to make sure
that we can move forward.

Change-Id: I9857803ef960efbd034f05985ec8fcc7272e2f70