Updated openstack/openstack Project: openstack-infra/project-config  b25deaf92ce7223d5d718373bcf9566b94e57a26
Add openstack/training-guides gerritbot for upstream-institute

Much of the work that we are doing in the
openstack-upstream-institute channel is making changes to
our slides in openstack/training-guides/doc/upstream-training .
It would be helpful to get notifications when patches are pushed
up and merged.

This patch adds that functionality to our channel.

Change-Id: I5bd19cd12b67460ebb64a5c426d69098b484baa5