Updated openstack/openstack Project: openstack/nova  559eb958f56e46cfa62eb9ac247ff357edb36308
Document and provide useful error message for volume-backed backup

The createBackup API does not support backing up volume-backed
instances. The error message the user gets is not useful about
why the request was invalid, and the API reference docs do not
mention the limitation. This change addresses both of those issues.

Change-Id: I04fd8ab4f8818d9d0ccccb6f6fcb34965b15b8f3
Partial-Bug: #1679314