Updated openstack/openstack Project: openstack/cinder-specs  5b540900c5d57312a649ab6e2ea87c9374419c0e
Allow an attached volume to be extended

The purpose of this spec is to change the behavior of the volume extend
API to allow an attached volume to be extended. This allows the
volume driver to extend the existing volume without first requiring
that the volume be detached and put back in the available state.

Nova will be asked to act on the size increase and to allow
the instance to have access to the added space after the volume
driver extends the volume.

blueprint: extend-attached-volume

Co-Authored-By: Gerald McBrearty <gfm@us.ibm.com>

Depends-On: I6622b154a7ebd5ebef37efedb23fc4b4401c052a
Change-Id: Id5b70ddac617a2fe0da9f2c2c0286a5ef61ef782