Updated openstack/openstack Project: openstack-infra/system-config  5f76abb55ffc7979e8afdbcf899b8888ae7aca8e
Move puppet service disablement to o_p::server

We don't need to stop the puppet service in single_use_slave, so as part
of emptying out openstack_project::template, move that resource
to openstack_project::server.

We still need to disable the service during the image build so add that
to the install_puppet.sh script.

Change-Id: I11db1b49f083c7a30e7908ba5a4a7df9d4033c9f