Updated openstack/openstack Project: openstack-infra/tripleo-ci  74f81e551cfbe00cf3457c8722b256089dac313a
Automatically read both voting and non-voting forms of job

In general we don't care about the voting or non-voting status of
a job on the cistatus page.  We just want to see whether a job is
passing or not.  There are also some jobs running in both voting
and non-voting form depending on the patch, which further complicates
things.

This change makes the job name list use only the voting form of the
names and handles the -nv form automatically if it is found.  In
addition to getting us all the relevant results for a job in one
column, this means we won't have to update the cistatus script
every time a job moves from non-voting to voting or vice versa.

Change-Id: I69294bf0e78d5e443e4b1c869991fc324c2d7c23