Updated openstack/openstack Project: openstack/tripleo-heat-templates  0991ed0dfb46194f4d41255fb87ee624f50028b8
Use conditionals for neutron and glance worker defaults

Using an empty string to signal that the default value in the puppet module
is to be used no longer seems to work, resulting in the puppet specified
defaults being overridden by empty string values. The impact on
configuration will differ depending on the actual configuration item, the
puppet code and the service, so it is just safer to omit the hieradata if
the user has not explicitly set a value.

Change-Id: Iefbc8f8669680e4f9d01db6b49543bfbe9b7661b
Closes-Bug: #1669452