Updated openstack/openstack Project: openstack-infra/project-config  73564ba282c1449028dcd36882064d644fc5b1df
Add mosquitto grafana dashboard

Now that we have stats reporting into statsd from the mosquitto broker
lets add a grafana dashboard to track that data. This is a first pass at
the dashboard I expect we'll have to update and refine it before we can
get useful metrics from it.

Change-Id: I1c4bc8fc8df544d95ec3711a065d88ec17df4078