Updated openstack/openstack Project: openstack-infra/project-config  35e4692647ed900b0047f9cb55614c69cf6fd46f
jenkins: scripts: run-tox.sh: Increase verbosity

Level 2 verbosity gives some useful output such as the PATH that's used
by tox and the start and finish time for each test which can be useful
when debugging tox related issues. Other than that, the output is very
similar to that of level 1.

Change-Id: I34a82bf45546e7b1b6265ecf8a7e9fd73a6cb3a9
Link: https://github.com/tox-dev/tox/search?utf8=%E2%9C%93&q=verbosity2&type=