Updated openstack/openstack Project: openstack/nova  60996ee9e28b89e2f4acb52ea97de3517fe73bfa
Update detach to use V3 Cinder API

Use the new Cinder V3 attachment delete method
during detach if the BDM has an attachment_id in it.
This will only be present in the BDM if/when the
new attachment_create API is called. Otherwise, we
revert to the old calls.

Edge cases are handled in separate patches.

Partially Implements: blueprint cinder-new-attach-apis

Co-Authored-By: Steve Noyes <steve.noyes@oracle.com>

Change-Id: I91b9a60268354ffbed86b1e7d173906cfd7b97bd