Updated openstack/openstack Project: openstack-infra/system-config  6143f0675f9c68d55da1ac09456c8b0df0a9b2b9
centos7: don't install EPEL

EPEL is not really helping us test, because upstream RDO (and hence
centos openstack) are designed to run without it.  In fact it seems to
actively get in the way.

Installing EPEL here is really a hang-over from the snapshot days when
we'd boot provider images without it.  We do need EPEL for bringup but
should have it default disabled so we whitelist use; see depends-on
which moves it into dib.

Depends-On: I2d060e048f9106292f9024efe0fd1d529d801a30
Change-Id: If6cf1bc300c6b59a8defb09fb3a3a1254e392a43