Updated openstack/openstack Project: openstack/nova  893ee87d3c40675383df286eac84a8093430c154
Tell people that the nova-cells man page is for cells v1

We have to continue scrubbing our documentation for anything
that generically refers to 'cells' since older documentation is
referring to cells v1, and we don't want people to confuse that
with cells v2. This simply provides a warning that cells v1 is
not recommended for new deployments and if you were looking for
help with cells v2 commands, we need to redirect you.

Change-Id: I3bd0332607bff169b5f496d17e4fc7d372ea991b