Updated openstack/openstack Project: openstack/diskimage-builder  803d40b0c61b360afe9850826467749d420115b6
Refactor block_device: passing command line parameters

The original approach was to pass each and every command
line parameter to the block device.  While the block device
functionality gets extended, this is not any longer practical.

Instead of passing in all the parameters separately this patch
collects these in a YAML file that is passed in to the block device
layer.

Change-Id: I9d07593a01441b62632234468ac25a982cf1a9f0
Signed-off-by: Andreas Florath <andreas@florath.net>