Updated openstack/openstack Project: openstack/training-guides  f4ef15ce74795e9d4b39e0fcfbf7862a533d055b
[upstream] Add links to and improve wording of communication slides

There was some odd wording in the communication slides.  This patch
improves the wording.

Also it adds a note that most teams put meeting agendas on the
team's wiki page.

Links are added to etherpad.openstack.org and to the wiki that
describes how to subscribe to mailing lists/describes our more
popular lists.

Change-Id: Ib9c5f06a873abe6a00e8d0fa8c700f8ad5f1ba4a