Updated openstack/openstack Project: openstack-infra/shade  837b5f406d66b527ab42aec8f84bf84067a5e984
Don't do all the network stuff in the rebuild poll

When we're polling for completeness during rebuild server, we do not
need to make all of the network calls to fill in network info.

If we wait for the server though, we do return the server to the user,
so fill in the info once at the end.

Change-Id: I78aaf995cb83a88fef9414c17b6bd7fca21fe3c3