Updated openstack/openstack Project: openstack/nova  0039231719d2a02c59e7cd76631e2ff03cc86b0d
Remove archaic reference to QEMU errors during post live migration

The compute manager should be virt-driver agnostic for the
most part. This reference to errors in post-live migration
for QEMU predates gerrit and goes back to a diablo commit
with no further explanation of what's going on with this
change, but it looks like it's what added live migration
way back when:

1518e66d0c4a4a72b6893cb117648f81dd3e3aff

It doesn't even make sense now, so remove it.

Change-Id: I144711f9e000c7e723f3b4e3dce56e3e922cf6e4