Updated openstack/openstack Project: openstack/tempest  8bb5abee3526a56fe7921b6327d3e0ecb87aa875
Remove duplication of skip_tracker

This commit removes of the duplication of skip_tracker in
tempest/lib/cmd and tools. And we already have 'skip-tracker' command.
It comes from tempest/lib/cmd/skip_tracker.py, so I think we don't need
to have duplicated code. So we should deprecate it in tools directory.

This commit also add a default test_path value for convenience and
compatibility.

Change-Id: I43a0a9d115e59c860714a47aea11017e8b58b5e7