Updated openstack/openstack Project: openstack/tripleo-heat-templates  f1f6b5dc7d698a36f04186856fb94b4115d121dc
Deploy ceilometer_auth_enabled to node containing keystone

This hiera key is used by keystone to create the ceilometer service
user. It works in CI cause keystone and the ceilometer services are in
the same node. However, this fails if keystone is deployed on a separate
note.

We should only deploy it in the nodes containing the keystone service
since it's only relevant to create the service user.

Change-Id: Ic0f02fe9a78a1fe14ac2b87197692fbd80c003b8
Closes-Bug: #1685828