Updated openstack/openstack Project: openstack/ironic  9558eb85e711065cf33eb1bc98f3eb227d1df533
Add a paragraph about image validation to Install Guide

When Nova is used with Ironic, Nova passes image information when a
new instance is requested. Because most of deploy drivers require the
image information, a node fails validation when it is enrolled. This
patch add a paragraph to Install guide so that an operator can ignore
this failure.

Change-Id: I2fd92c29b75de2891ac5556417b5ee81f8616a5a