Updated openstack/openstack Project: openstack/training-guides  c5ee52ca75744548a076969b2f8cd1b71509eafa
[upstream] Additional improvements for official-projects

In a discussion on IRC it was noted that we should move some of
the 'notes' information to the actual slides.  I made this change
for the project navigator so that it was clear that this was just
a list of some, not all, the projects.

I also don't know how I left the second bullet on the first slide
unedited.  I have now improved that wording as well to be clearer.

Change-Id: Iec8d3055ce43a01635c99892193a069f7fca83b7