Updated openstack/openstack Project: openstack/training-guides  86ce92eab80c8641b4bfeb33ad4910530a127b33
[upstream] Add core info to governance slides

While running through the slides with my team they were
surprised we didn't have anything about the core team members
along with the description of the Project Teams.  So,
adding mention of that here, and re-ordering the slides so that
we have the Project Team description immediately followed by
description of the PTL and Core Team Members.

Change-Id: I8ef76189394684ae273e31a514dd3ee0a85a6db2