Updated openstack/openstack Project: openstack/zaqar  cb1bde31ce7c8891dfcfdc6023a40ec9411bd89c
Fix swift messages posting when pooling

1. Currently, swift driver is always getting the swift queue controller
which is not correct when pooling is enabled.

2. When using pooling, all the queues will be saved in management
database, it works for database backend. But it doesn' work for swift,
because for database backend, the queue name is just a field of the
message record. For swift, we're using container as the "queue" to
save messages. So before saving messages, we need to make sure the
container of the queue has been created.

Closes-Bug: 1680678

Change-Id: I5195911d22d5935aa3de105fe735ac159276a025