Updated openstack/openstack Project: openstack/oslo.config  12fd34b963f6efa1bd8afd8ce06de33c90135f6d
add deprecation_reason to log message

When we log the warning that an option is deprecated for removal,
include the reason if we have one. Previously the deprecation reason was
only visible in the sample configuration file.

Add some tests for the log messages emitted when deprecated options are
used.

Change-Id: I5e309a3651041580fdf529ff31e18bbd90714f35
Signed-off-by: Doug Hellmann <doug@doughellmann.com>