Updated openstack/openstack Project: openstack/training-guides  b45321cf98143b16aefcc096aad8ec4336677af6
[upstream] Update Task Tracking Section

This patch greatly reduces the amount of lecture and focuses on the basics
of what new contributor need to know. It cleans up a lot of redundant
information as well. This patch adds exercises related to both launchpad
and storyboard both. It also renames the file since the slides are no
longer only about launchpad.

Change-Id: I170b26ba6d633f91c4d87a9c845271a93193670c