Updated openstack/openstack Project: openstack/horizon  3a8e7b67c637a22d4d99c348547cf22cb3959e40
Delete instance,volume,etc confirmation is missing name or ID

If click the name link and go into the details page for the instance, volume, etc.
And attempt to delete. It will pop up a message, but omit the name/id.
Need to add the property of 'data-display' in detail page.

Change-Id: I9c1966dc7942a27903237e2786edbcd2fbd936ea
Closes-Bug: #1680329