Updated openstack/openstack Project: openstack-infra/puppet-subunit2sql  4037e7b1ba93fbef0414d6b0adefd18e1076d910
Handle mqtt errors on failures before out_event is set

There are some conditions where we fail before out_event is defined. In
those cases the mqtt publish won't work and will raise an exception
because out_event is not defined. This commit addresses that by using
fields instead of out_event. (which is the original source for
out_event) Fields exists earlier and build_uuid always remains the same
across the 2 dicts. Additionally we define fields to an empty dict
outside of the try to ensure that in the off chance of a failure before
fields is defined we don't fail sending a failure on mqtt.

Change-Id: I546a2ee42ea0f67f35c5e132d51f239a32a5582e